+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - New subcommand `extract` allows for extraction of lines for the purpose of VO.
+- New subcommand `graph` allows for exporting a dot graph file of your dialogue.
 
 ### Changed